"Create function game"
The function game() takes the round of rock paper scissors
stablished at the function playRound() and puts on loop to be
played five times, as required by the project guide. It asks on a
prompt window which option the player chooses and displays on a
alert window if the player won, lost or tied the round. Also on a
alert window, shows the current score of the game.